+++ src/main/java/ch/spacebase/openclassic/server/network/handler/PlayerTeleportMessageHandler.java
@@ -15,7 +15,7 @@ public class PlayerTeleportMessageHandler extends MessageHandler<PlayerTeleportM
 		if(player == null || session == null) return;
 		if(session.getState() != State.GAME) return;
 		
-		Position to = new Position(player.getPosition().getLevel(), message.getX(), message.getY(), message.getZ(), message.getYaw(), message.getPitch());
+		Position to = new Position(player.getPosition().getLevel(), (float) message.getX(), (float) message.getY(), (float) message.getZ(), message.getYaw(), message.getPitch());
 		/* TODO: detect teleports properly if(!player.teleported && (to.getX() - player.getPosition().getX() >= 3 || to.getX() - player.getPosition().getX() <= -3 || to.getZ() - player.getPosition().getZ() >= 3 || to.getZ() - player.getPosition().getZ() <= -3)) {
 			PlayerRespawnEvent event = EventFactory.callEvent(new PlayerRespawnEvent(player, to));
 			if(event.isCancelled()) {
@@ -32,12 +32,12 @@ public class PlayerTeleportMessageHandler extends MessageHandler<PlayerTeleportM
 		to = event.getTo();
 		
 		if(event.isCancelled()) {
-			session.send(new PlayerTeleportMessage((byte) -1, event.getFrom().getX(), event.getFrom().getY(), event.getFrom().getZ(), event.getFrom().getYaw(), event.getFrom().getPitch()));
+			session.send(new PlayerTeleportMessage((byte) -1, event.getFrom().getX(), event.getFrom().getY(), event.getFrom().getZ(), (byte) event.getFrom().getYaw(), (byte) event.getFrom().getPitch()));
 			return;
 		}
 		
 		if(!to.equals(old)) {
-			session.send(new PlayerTeleportMessage((byte) -1, to.getX(), to.getY(), to.getZ(), to.getYaw(), to.getPitch()));
+			session.send(new PlayerTeleportMessage((byte) -1, to.getX(), to.getY(), to.getZ(), (byte) to.getYaw(), (byte) to.getPitch()));
 		}
 		
 		player.getPosition().setX(to.getX());
@@ -46,7 +46,7 @@ public class PlayerTeleportMessageHandler extends MessageHandler<PlayerTeleportM
 		player.getPosition().setYaw(to.getYaw());
 		player.getPosition().setPitch(to.getPitch());
 		
-		player.getPosition().getLevel().sendToAllExcept(player, new PlayerTeleportMessage(player.getPlayerId(), to.getX(), to.getY() + 0.59375, to.getZ(), to.getYaw(), to.getPitch()));
+		player.getPosition().getLevel().sendToAllExcept(player, new PlayerTeleportMessage(player.getPlayerId(), to.getX(), to.getY() + 0.59375, to.getZ(), (byte) to.getYaw(), (byte) to.getPitch()));
 	}
 
 }
